+++ src/pageEditor/tabs/dataPanelTabs.module.scss
@@ -35,11 +35,6 @@
   flex-direction: column;
 }
 
-.selectablePreviewContainer {
-  padding-left: 15px;
-  padding-right: 15px;
-}
-
 .tabPane {
   padding-top: 1rem;
 }
