+++ src/tsconfig.strictNullChecks.json
@@ -170,11 +170,7 @@
     "./components/formBuilder/RjsfTemplates.ts",
     "./components/formBuilder/TextAreaWidget.tsx",
     "./components/formBuilder/formBuilderTypes.ts",
-    "./components/formBuilder/preview/FormPreviewArrayField.tsx",
-    "./components/formBuilder/preview/FormPreviewBooleanField.tsx",
-    "./components/formBuilder/preview/FormPreviewFieldTemplate.tsx",
     "./components/formBuilder/preview/FormPreviewSchemaField.tsx",
-    "./components/formBuilder/preview/FormPreviewStringField.tsx",
     "./components/formBuilder/preview/ImageCropWidgetPreview.tsx",
     "./components/formBuilder/schemaFieldNames.ts",
     "./components/jsonTree/JsonTree.tsx",
